+++ react-native/storyly-react-native/RNStoryly.js
@@ -149,6 +149,7 @@ Storyly.propTypes = {
     storylyId: string.isRequired,
     storylySegments: arrayOf(string),
     storylyUserProperty: object,
+    storylyShareUrl: string,
     customParameter: string,
     storylyTestMode: bool,