+++ react-native/storyly-react-native/android/src/main/java/com/appsamurai/storyly/reactnative/STStorylyView.kt
@@ -135,7 +135,9 @@ class STStorylyView(context: Context) : FrameLayout(context) {
             storyMap.putInt("id", story.id)
             storyMap.putInt("index", story.index)
             storyMap.putString("title", story.title)
+            storyMap.putString("name", story.name)
             storyMap.putBoolean("seen", story.seen)
+            storyMap.putInt("currentTime", story.currentTime.toInt())
             storyMap.putMap("media", Arguments.createMap().also { storyMediaMap ->
                 storyMediaMap.putInt("type", story.media.type.ordinal)
                 storyMediaMap.putString("actionUrl", story.media.actionUrl)